added logic to output some variables in the main.js
