+++ src/index.html
@@ -7,8 +7,18 @@
   <header></header>
   <div class="container">
       
-        <button class="btn btn-success btn-lg" id="executeButton">Search</button>
+    <button class="btn btn-success btn-lg" id="executeButton">Search</button>
       
+    <div id="outputRentLow"></div>
+    <div id="outputRentHigh"></div>
+    <div id="outputListing1Address"></div>
+    <div id="outputListing1Price"></div>
+    <div id="outputListing1Bedrooms"></div>
+    <div id="outputListing1Bathrooms"></div>
+    <div id="outputListing1SquareFootage"></div>
+    <div id="outputListing1DaysOnMarket"></div>
+    <!-- test div -->
+    <div id="result"></div>
   </div>
 </body>
 </html>
